fixed alignment of c structs

+++ include/RoboCupGameControlData.h
@@ -83,7 +83,7 @@ struct TeamInfo
 struct RoboCupGameControlData
 {
   char header[4];               // header to identify the structure
-  uint8_t version;              // version of the data structure
+  uint16_t version;             // version of the data structure
   uint8_t packetNumber;         // number incremented with each packet sent (with wraparound)
   uint8_t playersPerTeam;       // The number of players on a team
   uint8_t gameType;             // type of the game (GAME_ROUNDROBIN, GAME_PLAYOFF)
